# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PESkinPro Backend API",
    "version": "1.0.0",
    "description": "API Documentation for PESkinPro E-Commerce Backend",
    "contact": {
      "name": "Developer"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Current Server (Relative)"
    },
    {
      "url": "http://localhost:5000/api/v1",
      "description": "API Server (Configured)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/address": {
      "post": {
        "summary": "Tambah Alamat Baru",
        "description": "Menyimpan alamat pengiriman baru untuk pelanggan.",
        "tags": [
          "User - Customer"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "recipient_name",
                  "phone_number",
                  "address_line",
                  "province_id",
                  "city_id",
                  "district_id"
                ],
                "properties": {
                  "recipient_name": {
                    "type": "string",
                    "example": "Budi Santoso"
                  },
                  "phone_number": {
                    "type": "string",
                    "example": "08123456789"
                  },
                  "address_line": {
                    "type": "string",
                    "description": "Detail alamat (Jalan, No Rumah)."
                  },
                  "province_id": {
                    "type": "integer"
                  },
                  "city_id": {
                    "type": "integer"
                  },
                  "district_id": {
                    "type": "integer"
                  },
                  "postal_code": {
                    "type": "string"
                  },
                  "is_default": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Alamat berhasil ditambahkan."
          }
        }
      }
    },
    "/addresses/{user_id}": {
      "get": {
        "summary": "Daftar Alamat Pelanggan",
        "description": "Mengambil semua daftar alamat yang dimiliki oleh seorang pengguna berdasarkan ID-nya.",
        "tags": [
          "User - Customer"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "UUID pengguna."
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar alamat berhasil ditemukan."
          }
        }
      }
    },
    "/address/{id}": {
      "put": {
        "summary": "Perbarui Alamat",
        "description": "Mengubah detail informasi pada alamat yang sudah ada.",
        "tags": [
          "User - Customer"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID unik alamat."
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "recipient_name": {
                    "type": "string"
                  },
                  "phone_number": {
                    "type": "string"
                  },
                  "address_line": {
                    "type": "string"
                  },
                  "province_id": {
                    "type": "integer"
                  },
                  "city_id": {
                    "type": "integer"
                  },
                  "district_id": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Alamat berhasil diperbarui."
          }
        }
      },
      "delete": {
        "summary": "Hapus Alamat",
        "description": "Menghapus alamat pengiriman dari daftar pelanggan.",
        "tags": [
          "User - Customer"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alamat berhasil dihapus."
          }
        }
      }
    },
    "/address/{id}/default": {
      "put": {
        "summary": "Setel Alamat Utama",
        "description": "Menjadikan salah satu alamat sebagai alamat pengiriman utama (default).",
        "tags": [
          "User - Customer"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Alamat berhasil disetel sebagai utama."
          }
        }
      }
    },
...
  },
...
}